renaming utils to sutils, and fixing all imports to resolve name conflict
