+++ libexec/python/Makefile.am
@@ -2,7 +2,7 @@ SUBDIRS = docker shub helpers
 
 scriptlibexecdir = $(libexecdir)/singularity/python
 
-dist_scriptlibexec_DATA = defaults.py __init__.py logman.py shell.py utils.py base.py templates.py
+dist_scriptlibexec_DATA = defaults.py __init__.py logman.py shell.py sutils.py base.py templates.py
 dist_scriptlibexec_SCRIPTS = import.py pull.py
 
 MAINTAINERCLEANFILES = Makefile.in
