+++ libexec/python/helpers/json/main.py
@@ -25,7 +25,7 @@ import sys
 import os
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), os.path.pardir)))
 
-from utils import (
+from sutils import (
     read_json,
     write_json
 )
