+++ libexec/python/sutils.py
