+++ libexec/python/tests/test_custom_cache.py
@@ -51,7 +51,7 @@ class TestCustomCache(TestCase):
         '''test_run_command tests sending a command to commandline
         using subprocess
         '''
-        print("Testing utils.get_cache with environment set")
+        print("Testing get_cache with environment set")
         from defaults import SINGULARITY_CACHE
         self.assertEqual(self.custom_cache,SINGULARITY_CACHE)   
         self.assertTrue(os.path.exists(SINGULARITY_CACHE))
